flow fixes

+++ src/components/ModalForm/ModalForm.js
@@ -66,24 +66,24 @@ class ModalForm extends React.Component<Props> {
 
   _onDelete = async (values: Values) => {
     const { onRequestClose, onDelete } = this.props;
+    if (!onDelete) return;
 
-    const res = await onDelete(values);
-
+    await onDelete(values);
     onRequestClose();
-
-    return res;
+    return;
   };
 
-  _onCancel = async (formState: any) => {
+  _onCancel = async (formState: FormStateAndHelpersAndActions) => {
     const { onRequestClose } = this.props;
 
     if (formState.dirty) {
       // TODO: use proper alert component, quit if user confirms
-      return !alert('not today, punk');
+      !alert('not today, punk');
+      return;
     }
 
-    const res = await onRequestClose();
-    return res;
+    onRequestClose();
+    return;
   };
 }
